+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/Layers/BubbleLayer.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Sources;
+using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
 namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
@@ -6,7 +7,7 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
     /// <summary>
     /// <inheritdoc cref="LayerType.Bubble"/>
     /// </summary>
-    public class BubbleLayer : LayerBase<BubbleLayerOptions, DataSource>
+    public class BubbleLayer : LayerBase<BubbleLayerOptions>
     {
         /// <summary>
         /// <inheritdoc cref="LayerType"/>