+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/Layers/HeatMapLayer.cs
@@ -1,9 +1,11 @@
-﻿namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
+﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Sources;
+
+namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
 {
     /// <summary>
     /// <inheritdoc cref="LayerType.HeatMap"/>
     /// </summary>
-    public class HeatMapLayer : LayerBase<HeatMapLayerOptions, DataSource>
+    public class HeatMapLayer : LayerBase<HeatMapLayerOptions>
     {
         /// <summary>
         /// <inheritdoc cref="LayerType"/>