+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/Layers/PolygonLayer.cs
@@ -1,9 +1,11 @@
-﻿namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
+﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Sources;
+
+namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
 {
     /// <summary>
     /// <inheritdoc cref="LayerType.Polygon"/>
     /// </summary>
-    public class PolygonLayer : LayerBase<PolygonLayerOptions, DataSource>
+    public class PolygonLayer : LayerBase<PolygonLayerOptions>
     {
         /// <summary>
         /// <inheritdoc cref="LayerType"/>