+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/Layers/TileLayer.cs
@@ -1,11 +1,12 @@
-﻿using Marqdouj.DotNet.Web.JsInterop.GeoJson;
+﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Sources;
+using Marqdouj.DotNet.Web.JsInterop.GeoJson;
 
 namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Layers
 {
     /// <summary>
     /// <inheritdoc cref="LayerType.Tile"/>
     /// </summary>
-    public class TileLayer : LayerBase<TileLayerOptions, DataSource>
+    public class TileLayer : LayerBase<TileLayerOptions>
     {
         /// <summary>
         /// <inheritdoc cref="LayerType"/>